Schedule text with match-ups and team members combined
Reuse code via `!schedule` command.
